[feat] update files

+++ scripts/clean.sh
@@ -4,7 +4,10 @@ FILES=$(find . -type f | grep -Ei '\.[a-zA-Z0-9]+\~$')
 
 while IFS= read -r line; do
   FILE="$line"
-  echo "$FILE"
-  rm "$FILE"
+  if [ -f "$FILE" ]; then
+    echo "$FILE"
+    rm "$FILE"
+
+  fi
 
 done <<< "$FILES"
